requirements.txt
Flask
gunicorn
requests
newsapi-python
google-cloud-storage